requirements.txt
streamlit==1.28.0
pandas==2.0.3
plotly==5.15.0
numpy==1.24.3
openai==0.28.0
requests==2.31.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
scipy==1.11.0
scikit-learn==1.3.2
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
gspread==5.12.0
oauth2client==4.1.3